Move validateEmail/validateUrl to src/shared/validation.ts
Both functions were duplicated between sanitize.ts (main) and
contactValidation.ts (renderer). Consolidate into a shared module
accessible to both sides; sanitize.ts re-exports them so http-server.ts
and tests are unaffected.
