+++ src/main/sanitize.ts
@@ -1,14 +1,9 @@
 import { parsePhoneNumberFromString } from 'libphonenumber-js';
 import type { CountryCode } from 'libphonenumber-js';
+export { validateEmail, validateUrl } from '@shared/validation';
 
 export function normalizeEmail(raw: string): string {
   return raw.trim().toLowerCase();
-}
-
-export function validateEmail(raw: string): boolean {
-  const trimmed = raw.trim();
-  if (!trimmed) return false;
-  return /^[^\s@]+@[^\s@]+\.[^\s@]{2,}$/.test(trimmed);
 }
 
 export function normalizePhone(raw: string, defaultCountry: string = 'US'): string | null {
@@ -20,17 +15,6 @@ export function normalizePhone(raw: string, defaultCountry: string = 'US'): stri
     return parsed.ext ? `${base} ext. ${parsed.ext}` : base;
   }
   return null;
-}
-
-export function validateUrl(raw: string): boolean {
-  const trimmed = raw.trim();
-  if (!trimmed) return false;
-  try {
-    const url = new URL(trimmed);
-    return url.protocol === 'https:' || url.protocol === 'http:';
-  } catch {
-    return false;
-  }
 }
 
 export function detectLinkType(url: string): 'linkedin' | 'x' | 'instagram' | 'facebook' | 'website' {
